# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vehicle Transport Automation",
    "description": "Control Panel for Email-to-ClickUp Pipeline with ML Training Support",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/settings/export-targets": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Export Targets",
        "description": "Update which export targets are enabled.",
        "operationId": "update_export_targets_api_settings_export_targets_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportTargets"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/sheets": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Sheets Settings",
        "description": "Update Google Sheets settings.",
        "operationId": "update_sheets_settings_api_settings_sheets_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SheetsSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/clickup": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Clickup Settings",
        "description": "Update ClickUp settings.",
        "operationId": "update_clickup_settings_api_settings_clickup_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClickUpSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/cd": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Cd Settings",
        "description": "Update Central Dispatch settings.",
        "operationId": "update_cd_settings_api_settings_cd_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CDSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/email": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Email Settings",
        "description": "Update email ingestion settings.",
        "operationId": "update_email_settings_api_settings_email_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/warehouses": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get Warehouses",
        "description": "Get all configured warehouses.",
        "operationId": "get_warehouses_api_settings_warehouses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WarehouseConfig"
                  },
                  "type": "array",
                  "title": "Response Get Warehouses Api Settings Warehouses Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Warehouses",
        "description": "Update warehouse list.",
        "operationId": "update_warehouses_api_settings_warehouses_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/WarehouseConfig"
                },
                "type": "array",
                "title": "Warehouses"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Add Warehouse",
        "description": "Add a new warehouse.",
        "operationId": "add_warehouse_api_settings_warehouses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/warehouses/{warehouse_id}": {
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Delete Warehouse",
        "description": "Delete a warehouse by ID.",
        "operationId": "delete_warehouse_api_settings_warehouses__warehouse_id__delete",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CDSettings": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "username": {
            "type": "string",
            "title": "Username",
            "default": ""
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": ""
          },
          "shipper_id": {
            "type": "string",
            "title": "Shipper Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "CDSettings",
        "description": "Central Dispatch settings."
      },
      "ClickUpSettings": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "api_token": {
            "type": "string",
            "title": "Api Token",
            "default": ""
          },
          "list_id": {
            "type": "string",
            "title": "List Id",
            "default": ""
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ClickUpSettings",
        "description": "ClickUp settings."
      },
...
      "EmailSettings": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "imap_server": {
            "type": "string",
            "title": "Imap Server",
            "default": ""
          },
          "imap_port": {
            "type": "integer",
            "title": "Imap Port",
            "default": 993
          },
          "email_address": {
            "type": "string",
            "title": "Email Address",
            "default": ""
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": ""
          }
        },
        "type": "object",
        "title": "EmailSettings",
        "description": "Email ingestion settings."
      },
...
      "ExportTargets": {
        "properties": {
          "sheets": {
            "type": "boolean",
            "title": "Sheets",
            "default": false
          },
          "clickup": {
            "type": "boolean",
            "title": "Clickup",
            "default": false
          },
          "cd": {
            "type": "boolean",
            "title": "Cd",
            "default": false
          }
        },
        "type": "object",
        "title": "ExportTargets",
        "description": "Export target configuration."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SheetsSettings": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "spreadsheet_id": {
            "type": "string",
            "title": "Spreadsheet Id",
            "default": ""
          },
          "sheet_name": {
            "type": "string",
            "title": "Sheet Name",
            "default": "Pickups"
          },
          "credentials_file": {
            "type": "string",
            "title": "Credentials File",
            "default": "config/sheets_credentials.json"
          }
        },
        "type": "object",
        "title": "SheetsSettings",
        "description": "Google Sheets settings."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WarehouseConfig": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "zip_code": {
            "type": "string",
            "title": "Zip Code"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "city",
          "state",
          "zip_code"
        ],
        "title": "WarehouseConfig",
        "description": "Warehouse configuration."
      },
...
    }
  }
}